trying with pinia store
