+++ package.json
@@ -28,6 +28,7 @@
     "p-debounce": "^4.0.0",
     "p-map": "^6.0.0",
     "papaparse": "^5.4.1",
+    "pinia": "^2.1.7",
     "popper.js": "^1.16.1",
     "v-tooltip": "^2.0.3",
     "vue": "^3.1.0",
